Remove Chrome and Firefox from travis-ci config
This mostly has to do with the structure of the app/test suite. Tests create
minimal dom elements for tests, but when popup.js loaded, it makes assumptions
about the structure of the HTML. Since there are dom elements missing,
exceptions are thrown. The Chrome and Firefox versions on travis-ci behave
differently than the more modern versions used in development. These versions
log the error, but do not exit != 0.